add interrupt handler logic for machine, add memory segmentation

+++ src/int.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 #[derive(Debug)]
 pub struct InterruptController {
-    interrupts: HashMap<u8, u8>,
+    interrupts: HashMap<u16, u8>,
 }
 
 impl InterruptController {
@@ -12,12 +12,12 @@ impl InterruptController {
         }
     }
 
-    pub fn int(&mut self, dev: u8, int: u8) -> anyhow::Result<()> {
+    pub fn int(&mut self, dev: u8, int: u16) -> anyhow::Result<()> {
         self.interrupts.insert(int, dev);
         Ok(())
     }
 
-    pub fn int_ack(&mut self, int: u8) -> anyhow::Result<()> {
+    pub fn int_ack(&mut self, int: u16) -> anyhow::Result<()> {
         self.interrupts.remove(&int);
         Ok(())
     }
